Clean up MSVC build.  Fixes #1619.  Fixes #3074.
git-svn-id: http://svn.boost.org/svn/boost/trunk@62854 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ libs/fusion/test/sequence/adapt_assoc_class.cpp
@@ -98,14 +98,14 @@ main()
     {
         // conversion from ns::point to vector
         ns::point p(5, 3);
-        boost::fusion::vector<int, short> v(p);
+        boost::fusion::vector<int, long> v(p);
         v = p;
     }
 
     {
         // conversion from ns::point to list
         ns::point p(5, 3);
-        boost::fusion::list<int, short> l(p);
+        boost::fusion::list<int, long> l(p);
         l = p;
     }
