+++ libs/fusion/test/sequence/adapt_assoc_class_named.cpp
@@ -102,7 +102,7 @@ main()
         // conversion from adapted::point to vector
         ns::point basep(5, 3);
         adapted::point p(basep);
-        boost::fusion::vector<int, short> v(p);
+        boost::fusion::vector<int, long> v(p);
         v = p;
     }
 
@@ -110,7 +110,7 @@ main()
         // conversion from adated::point to list
         ns::point basep(5, 3);
         adapted::point p(basep);
-        boost::fusion::list<int, short> l(p);
+        boost::fusion::list<int, long> l(p);
         l = p;
     }
