+++ libs/fusion/test/sequence/adapt_class.cpp
@@ -106,14 +106,14 @@ main()
     {
         // conversion from ns::point to vector
         ns::point p(5, 3);
-        fusion::vector<int, short> v(p);
+        fusion::vector<int, long> v(p);
         v = p;
     }
 
     {
         // conversion from ns::point to list
         ns::point p(5, 3);
-        fusion::list<int, short> l(p);
+        fusion::list<int, long> l(p);
         l = p;
     }
