+++ libs/fusion/test/sequence/adapt_class_named.cpp
@@ -112,7 +112,7 @@ main()
         ns::point basep(5, 3);
         adapted::point p(basep);
 
-        fusion::vector<int, short> v(p);
+        fusion::vector<int, long> v(p);
         v = p;
     }
 
@@ -121,7 +121,7 @@ main()
         ns::point basep(5, 3);
         adapted::point p(basep);
 
-        fusion::list<int, short> l(p);
+        fusion::list<int, long> l(p);
         l = p;
     }
