+++ libs/fusion/test/sequence/adapt_struct_named.cpp
@@ -103,7 +103,7 @@ main()
         // conversion from adapted::point to vector
         ns::point basep = {5, 3};
         adapted::point p(basep);
-        fusion::vector<int, short> v(p);
+        fusion::vector<int, long> v(p);
         v = p;
     }
 
@@ -111,7 +111,7 @@ main()
         // conversion from adapted::point to list
         ns::point basep = {5, 3};
         adapted::point p(basep);
-        fusion::list<int, short> l(p);
+        fusion::list<int, long> l(p);
         l = p;
     }
