+++ libs/fusion/test/sequence/define_assoc_struct.cpp
@@ -75,14 +75,14 @@ main()
     {
         // conversion from ns::point to vector
         ns::point p(5, 3);
-        vector<int, short> v(p);
+        vector<int, long> v(p);
         v = p;
     }
 
     {
         // conversion from ns::point to list
         ns::point p(5, 3);
-        list<int, short> l(p);
+        list<int, long> l(p);
         l = p;
     }
